refactor: Move the Parser trait into the parser module
BREAKING CHANGE

`Parser` is no longe exported from `primitives`.
Change `use combine::primitives::Parser` to `use combine::Parser`
